pyproject.toml
[project]
name = "crashlens"
version = "0.1.0"
description = "CLI to detect GPT token waste from Langfuse logs"
authors = [
    {name = "Aditya Singh", email = "coding103856@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "click >=8.2.1,<9.0.0",
    "pyyaml >=6.0.2,<7.0.0",
    "jinja2 >=3.1.6,<4.0.0",
    "rich >=14.0.0,<15.0.0",
    "orjson >=3.10.18,<4.0.0"
]

[project.scripts]
crashlens = "crashlens.cli:cli"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
